refactor: Layer 생성 위치 변경
