--- src/main/java/model/StepExistenceGenerator.java
@@ -1,14 +1,0 @@
-package model;
-
-import java.util.Random;
-
-public class StepExistenceGenerator {
-    private static final Random random = new Random();
-
-    public static boolean generate(boolean priorExistence) {
-        if (priorExistence) {
-            return false;
-        }
-        return random.nextBoolean();
-    }
-}
